test that all nodes traced are leaves
kernel sampling can add a kernel call path as a descendant of a path that
has already been recorded in a trace. test that traced nodes continue to
appear as leaves. hpcrun splits a traced node with machine code address
A and hangs the kernel path off a sibling that has a machine code address
of A+1.